backend/requirements.txt
# CareerPath AI Backend Requirements
boto3>=1.34.0
pytest>=7.4.0
pydantic>=2.5.0
PyPDF2>=3.0.0
python-docx>=1.1.0
